refactor: admin 페이지 파일명 통일 #107
