+++ src/main/java/com/goorm/roomflow/domain/notice/mapper/NoticeMapper.java
@@ -1,7 +1,7 @@
 package com.goorm.roomflow.domain.notice.mapper;
 
-import com.goorm.roomflow.domain.notice.dto.response.NoticeAdminDetailRes;
-import com.goorm.roomflow.domain.notice.dto.response.NoticeAdminRes;
+import com.goorm.roomflow.domain.notice.dto.response.AdminNoticeDetailRes;
+import com.goorm.roomflow.domain.notice.dto.response.AdminNoticeRes;
 import com.goorm.roomflow.domain.notice.dto.response.NoticeDetailRes;
 import com.goorm.roomflow.domain.notice.dto.response.NoticeRes;
 import com.goorm.roomflow.domain.notice.entity.Notice;
@@ -11,11 +11,11 @@ import org.mapstruct.Mapping;
 @Mapper(componentModel = "spring")
 public interface NoticeMapper {
 
-    NoticeAdminRes toNoticeAdminRes(Notice notice);
+    AdminNoticeRes toNoticeAdminRes(Notice notice);
 
     @Mapping(target = "createdByName", expression = "java(getUserName(notice.getCreatedBy()))")
     @Mapping(target = "updatedByName", expression = "java(getUserName(notice.getUpdatedBy()))")
-    NoticeAdminDetailRes toNoticeAdminDetailRes(Notice notice);
+    AdminNoticeDetailRes toNoticeAdminDetailRes(Notice notice);
 
     NoticeRes toNoticeRes(Notice notice);
     NoticeDetailRes toNoticeDetailRes(Notice notice);
