+++ src/main/java/com/goorm/roomflow/domain/notice/service/AdminNoticeService.java
@@ -1,14 +1,14 @@
 package com.goorm.roomflow.domain.notice.service;
 
 import com.goorm.roomflow.domain.notice.dto.request.NoticeReq;
-import com.goorm.roomflow.domain.notice.dto.response.NoticeAdminDetailRes;
-import com.goorm.roomflow.domain.notice.dto.response.NoticeAdminRes;
+import com.goorm.roomflow.domain.notice.dto.response.AdminNoticeDetailRes;
+import com.goorm.roomflow.domain.notice.dto.response.AdminNoticeRes;
 import org.springframework.data.domain.Page;
 
 public interface AdminNoticeService {
 
-    Page<NoticeAdminRes> readNoticeList(int page, int size);
-    NoticeAdminDetailRes readNoticeDetail(Long noticeId);
+    Page<AdminNoticeRes> readNoticeList(int page, int size);
+    AdminNoticeDetailRes readNoticeDetail(Long noticeId);
 
     void createNotice(Long userId, NoticeReq noticeReq);
     void modifyNotice(Long userId, Long noticeId, NoticeReq noticeReq);
