+++ src/main/java/com/goorm/roomflow/domain/reservation/dto/request/AdminReservationPolicyUpdateReq.java
@@ -1,6 +1,6 @@
 package com.goorm.roomflow.domain.reservation.dto.request;
 
-public record ReservationPolicyUpdateReq(
+public record AdminReservationPolicyUpdateReq(
         String policyValue
 ) {
 }
