+++ src/main/java/com/goorm/roomflow/domain/reservation/service/AdminReservationPolicyService.java
@@ -1,12 +1,12 @@
 package com.goorm.roomflow.domain.reservation.service;
 
-import com.goorm.roomflow.domain.reservation.dto.request.ReservationPolicyUpdateReq;
-import com.goorm.roomflow.domain.reservation.dto.response.ReservationPolicyAdminRes;
+import com.goorm.roomflow.domain.reservation.dto.request.AdminReservationPolicyUpdateReq;
+import com.goorm.roomflow.domain.reservation.dto.response.AdminReservationPolicyRes;
 
 import java.util.List;
 
 public interface AdminReservationPolicyService {
 
-    List<ReservationPolicyAdminRes> getPolicyList();
-    void updatePolicy(Long policyId, ReservationPolicyUpdateReq policyUpdateReq);
+    List<AdminReservationPolicyRes> getPolicyList();
+    void updatePolicy(Long policyId, AdminReservationPolicyUpdateReq policyUpdateReq);
 }
