+++ src/main/java/com/goorm/roomflow/domain/room/controller/AdminRoomSlotController.java
@@ -1,13 +1,10 @@
 package com.goorm.roomflow.domain.room.controller;
 
-import com.goorm.roomflow.domain.room.dto.request.RoomSlotDateStatusReq;
-import com.goorm.roomflow.domain.room.dto.request.RoomSlotGenerateReq;
-import com.goorm.roomflow.domain.room.dto.request.RoomSlotStatusReq;
-import com.goorm.roomflow.domain.room.dto.response.MeetingRoomAdminRes;
-import com.goorm.roomflow.domain.room.dto.response.RoomSlotAdminPageRes;
-import com.goorm.roomflow.domain.room.dto.response.RoomSlotAdminRes;
-import com.goorm.roomflow.domain.room.service.MeetingRoomService;
-import com.goorm.roomflow.domain.room.service.RoomSlotService;
+import com.goorm.roomflow.domain.room.dto.request.AdminRoomSlotDateStatusReq;
+import com.goorm.roomflow.domain.room.dto.request.AdminRoomSlotGenerateReq;
+import com.goorm.roomflow.domain.room.dto.request.AdminRoomSlotStatusReq;
+import com.goorm.roomflow.domain.room.dto.response.AdminRoomSlotPageRes;
+import com.goorm.roomflow.domain.room.service.AdminRoomSlotService;
 import com.goorm.roomflow.global.exception.BusinessException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.format.annotation.DateTimeFormat;
@@ -17,15 +14,13 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.time.LocalDate;
-import java.util.List;
 
 @Controller
 @RequiredArgsConstructor
 @RequestMapping("/admin/rooms/{roomId}/slots")
 public class AdminRoomSlotController {
 
-    private final MeetingRoomService meetingRoomService;
-    private final RoomSlotService roomSlotService;
+    private final AdminRoomSlotService adminRoomSlotService;
 
     /**
      * 특정 회의실의 특정 날짜 슬롯 목록 조회
@@ -39,14 +34,14 @@ public class AdminRoomSlotController {
     ) {
         LocalDate targetDate = (date != null) ? date : LocalDate.now();
 
-        RoomSlotAdminPageRes page = roomSlotService.getAdminRoomSlots(roomId, targetDate);
+        AdminRoomSlotPageRes page = adminRoomSlotService.getAdminRoomSlots(roomId, targetDate);
 
         model.addAttribute("roomId", roomId);
         model.addAttribute("selectedDate", targetDate);
         model.addAttribute("page", page);
 
-        model.addAttribute("generateForm", new RoomSlotGenerateReq(targetDate));
-        model.addAttribute("dateStatusForm", new RoomSlotDateStatusReq(targetDate, true));
+        model.addAttribute("generateForm", new AdminRoomSlotGenerateReq(targetDate));
+        model.addAttribute("dateStatusForm", new AdminRoomSlotDateStatusReq(targetDate, true));
 
         return "admin/system/roomslot-list";
     }
@@ -57,11 +52,11 @@ public class AdminRoomSlotController {
     @PostMapping("/generate")
     public String generateRoomSlots(
             @PathVariable Long roomId,
-            @ModelAttribute("generateForm") RoomSlotGenerateReq request,
+            @ModelAttribute("generateForm") AdminRoomSlotGenerateReq request,
             RedirectAttributes redirectAttributes
     ) {
         try {
-            roomSlotService.generateRoomSlots(roomId, request.date());
+            adminRoomSlotService.generateRoomSlots(roomId, request.date());
 
             redirectAttributes.addFlashAttribute("message", "슬롯이 생성되었습니다.");
             redirectAttributes.addFlashAttribute("alertType", "success");
@@ -80,11 +75,11 @@ public class AdminRoomSlotController {
     public String changeRoomSlotStatus(
             @PathVariable Long roomId,
             @PathVariable Long roomSlotId,
-            @ModelAttribute RoomSlotStatusReq request,
+            @ModelAttribute AdminRoomSlotStatusReq request,
             RedirectAttributes redirectAttributes
     ) {
         try {
-            roomSlotService.changeRoomSlotStatus(roomId, roomSlotId, request.active());
+            adminRoomSlotService.changeRoomSlotStatus(roomId, roomSlotId, request.active());
 
             redirectAttributes.addFlashAttribute("message", "슬롯 상태가 변경되었습니다.");
             redirectAttributes.addFlashAttribute("alertType", "success");
@@ -102,11 +97,11 @@ public class AdminRoomSlotController {
     @PostMapping("/date-status")
     public String changeDateRoomSlotStatus(
             @PathVariable Long roomId,
-            @ModelAttribute("dateStatusForm") RoomSlotDateStatusReq request,
+            @ModelAttribute("dateStatusForm") AdminRoomSlotDateStatusReq request,
             RedirectAttributes redirectAttributes
     ) {
         try {
-            roomSlotService.changeDateRoomSlotStatus(roomId, request.date(), request.active());
+            adminRoomSlotService.changeDateRoomSlotStatus(roomId, request.date(), request.active());
 
             redirectAttributes.addFlashAttribute("message", "해당 날짜 슬롯 상태가 변경되었습니다.");
             redirectAttributes.addFlashAttribute("alertType", "success");
