+++ src/main/java/com/goorm/roomflow/domain/room/dto/response/AdminRoomSlotPageRes.java
@@ -1,0 +1,9 @@
+package com.goorm.roomflow.domain.room.dto.response;
+
+import java.util.List;
+
+public record AdminRoomSlotPageRes(
+        AdminRoomSlotSummaryRes summary,
+        List<AdminRoomSlotRes> slots
+) {
+}
