+++ src/main/java/com/goorm/roomflow/domain/room/dto/response/AdminRoomSlotSummaryRes.java
@@ -4,7 +4,7 @@ import com.goorm.roomflow.domain.room.entity.RoomStatus;
 
 import java.time.LocalDate;
 
-public record RoomSlotSummaryRes(
+public record AdminRoomSlotSummaryRes(
         Long roomId,
         String roomName,
         RoomStatus roomStatus,
