--- src/main/java/com/goorm/roomflow/domain/room/dto/response/RoomSlotAdminPageRes.java
@@ -1,9 +1,0 @@
-package com.goorm.roomflow.domain.room.dto.response;
-
-import java.util.List;
-
-public record RoomSlotAdminPageRes(
-        RoomSlotSummaryRes summary,
-        List<RoomSlotAdminRes> slots
-) {
-}
