+++ src/main/java/com/goorm/roomflow/domain/room/repository/RoomSlotQueryRepository.java
@@ -1,7 +1,7 @@
 package com.goorm.roomflow.domain.room.repository;
 
-import com.goorm.roomflow.domain.room.dto.response.RoomSlotAdminRes;
-import com.goorm.roomflow.domain.room.dto.response.RoomSlotSummaryRes;
+import com.goorm.roomflow.domain.room.dto.response.AdminRoomSlotRes;
+import com.goorm.roomflow.domain.room.dto.response.AdminRoomSlotSummaryRes;
 import com.goorm.roomflow.domain.room.entity.MeetingRoom;
 
 import java.time.LocalDate;
@@ -9,8 +9,8 @@ import java.util.List;
 
 public interface RoomSlotQueryRepository {
 
-    List<RoomSlotAdminRes> findAdminRoomSlots(Long roomId, LocalDate date);
-    RoomSlotSummaryRes findRoomSlotSummary(MeetingRoom meetingRoom, LocalDate selectedDate);
+    List<AdminRoomSlotRes> findAdminRoomSlots(Long roomId, LocalDate date);
+    AdminRoomSlotSummaryRes findRoomSlotSummary(MeetingRoom meetingRoom, LocalDate selectedDate);
     boolean existsValidReservationByRoomSlotId(Long roomSlotId);
     boolean existsValidReservationByRoomIdAndDate(Long roomId, LocalDate date);
 }
