+++ src/main/java/com/goorm/roomflow/domain/room/repository/RoomSlotRepository.java
@@ -1,17 +1,13 @@
 package com.goorm.roomflow.domain.room.repository;
 
-import com.goorm.roomflow.domain.room.dto.response.RoomSlotAdminRes;
 import com.goorm.roomflow.domain.room.entity.MeetingRoom;
 import com.goorm.roomflow.domain.room.entity.RoomSlot;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
 
 public interface RoomSlotRepository extends JpaRepository<RoomSlot, Long> {
     List<RoomSlot> findBySlotStartAtGreaterThanEqualAndSlotStartAtLessThanOrderBySlotStartAtAsc(
